refactor: rename go 'ipfs' module to 'core'
