+++ go/bind/core/dnsconfig.go
@@ -1,7 +1,7 @@
 // from: https://gist.github.com/cs8425/107e01a0652f1f1f6e033b5b68364b5e
 
 //nolint
-package ipfs
+package core
 
 import (
 	_ "net"
