+++ go/bind/core/dnsconfig_empty.s
