+++ go/bind/core/init.go
@@ -1,4 +1,4 @@
-package ipfs
+package core
 
 import (
 	"encoding/base64"
