+++ go/bind/core/node.go
@@ -3,7 +3,7 @@
 // This package intend to only be use with gomobile bind directly if you
 // want to use it in your own gomobile project, you may want to use host/node package directly
 
-package ipfs
+package core
 
 // Main API exposed to the ios/android
