+++ go/bind/core/node_test.go
@@ -1,4 +1,4 @@
-package ipfs
+package core
 
 import (
 	"context"
