+++ go/bind/core/shell.go
@@ -1,4 +1,4 @@
-package ipfs
+package core
 
 import (
 	"context"
