+++ go/bind/core/sockmanager_test.go
@@ -1,4 +1,4 @@
-package ipfs
+package core
 
 import (
 	"io/ioutil"
