+++ go/bind/core/testing.go
@@ -1,4 +1,4 @@
-package ipfs
+package core
 
 import (
 	"io/ioutil"
